deploy order items validating correctly

+++ server/src/services/OrdersService.js
@@ -1,9 +1,13 @@
 import { dbContext } from "../db/DbContext.js"
-import { BadRequest } from "../utils/Errors.js"
+import { BadRequest, Forbidden } from "../utils/Errors.js"
+import { orderItemsService } from "./OrderItemsService.js"
+import { restaurantsService } from "./RestaurantsService.js"
 
 class OrdersService {
   async decimateOrder(userId, orderId) {
     const order = await dbContext.Orders.findById(orderId)
+    const restaurant = await restaurantsService.getRestaurantById(order.restaurantId)
+    if(userId != order.accountId ||  restaurant.creatorId != userId)throw new Forbidden('You didnt make this / dont own this restaurant')
     if (order.completed || order.isCancelled == true) throw new BadRequest('Already Completed')
     if (order.placed == true) order.isCancelled = true
     await order.deleteOne()
@@ -27,7 +31,7 @@ class OrdersService {
     const orders = await dbContext.Orders.find({ accountId: accountId }).populate('restaurant')
     return orders
   }
-  
+
   async getSpecificOrder(orderId) {
     const order = await dbContext.Orders.findById(orderId).populate('profile restaurant')
     return order
@@ -37,6 +41,15 @@ class OrdersService {
     const order = await dbContext.Orders.create(orderData)
     await order.populate('profile restaurant')
     return order
+  }
+
+  async realDeleteOrder(userId, orderId) {
+    const order = await dbContext.Orders.findById(orderId)
+    const restaurant = await restaurantsService.getRestaurantById(order.restaurantId)
+    if(userId != order.accountId && restaurant.creatorId != userId)throw new Forbidden('You didnt make this / dont own this restaurant')
+    await orderItemsService.deleteOrderItems(orderId)
+    await order.deleteOne()
+    // THIS DOES NOT STAY HERE
   }
 
 }
